tlt-1870: dashboard - enables and moves lti launch tests into dashboard app
